IBBS Backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.21.1
SQLAlchemy>=1.4
asyncpg>=0.26.0
alembic>=1.11.1
pydantic>=1.10.7
python-dotenv>=0.21.0
redis>=4.5.0
celery[redis]>=5.3.0
rq>=1.14.0
typing-extensions
prometheus-client>=0.15.0
Jinja2>=3.1.2
python-json-logger>=2.0.4
sentry-sdk>=1.21.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
